tutorials/tracing/teaching-tracingManualMLsys/src/teaching_tracingmanual/preprocessor/requirements.txt
flask==2.3.3
pillow==10.0.1
requests==2.31.0
pika
psutil
prometheus-client
paho-mqtt==1.6.1
pandas
dnspython==2.2.1
devtools
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
